+++ .gitignore
@@ -1,4 +1,24 @@
-/node_modules/
+# Logs
+logs
+*.log
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
+pnpm-debug.log*
+lerna-debug.log*
 
+node_modules
+dist
+dist-ssr
+*.local
+
+# Editor directories and files
+.vscode/*
+!.vscode/extensions.json
+.idea
 .DS_Store
-template-master.zip
+*.suo
+*.ntvs*
+*.njsproj
+*.sln
+*.sw?
