+++ scripts_old/setupTypeScript.js
