+++ src_old/translate.js
